+++ lib/ui/inspector_ui.dart
@@ -1,48 +1,93 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_network_inspector/net_inspect.dart';
+import 'package:flutter_network_inspector/utils/constants.dart';
 import 'package:flutter_network_inspector/utils/utils.dart';
 
 class InspectorUI extends StatelessWidget {
-  const InspectorUI({super.key});
+  const InspectorUI({
+    super.key,
+    this.title = 'Activity',
+  });
+
+  final String title;
 
   @override
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
         elevation: 3,
-        title: const Text('Activity'),
+        title: Text(title),
       ),
       body: ValueListenableBuilder(
         valueListenable: FNICLient.inspectorNotifierList,
         builder: (context, inspectorResults, _) {
           return ListView.separated(
-            separatorBuilder: (_, __) => const Divider(color: Colors.black12),
+            separatorBuilder: (_, __) => const Divider(
+              color: Colors.black12,
+              height: 0.3,
+              thickness: .5,
+            ),
             itemCount: inspectorResults.length,
             itemBuilder: (context, index) {
               final data = inspectorResults[index];
               final success = isSuccess(data.statusCode ?? 0);
               return ListTile(
                 dense: true,
-                enableFeedback: true,
-                onTap: () {
+                visualDensity: VisualDensity.compact,
+                contentPadding: const EdgeInsets.fromLTRB(10, 3, 10, 3),
+                isThreeLine: false,
+                onTap: () async {
                   Navigator.of(context).pushNamed('/details', arguments: data);
                 },
-                leading: _LeadingStatusIncdicatorBlock(
-                  success: success,
-                  method: data.baseRequest?.method ?? ' <null>',
+                leading: Column(
+                  crossAxisAlignment: CrossAxisAlignment.start,
+                  mainAxisAlignment: MainAxisAlignment.center,
+                  children: [
+                    Text(
+                      '${data.url?.scheme.toUpperCase()}',
+                      maxLines: 1,
+                      style: const TextStyle(
+                        color: Colors.brown,
+                        fontWeight: FontWeight.bold,
+                      ),
+                    ),
+                    Text(
+                      data.baseRequest?.method ?? noValue,
+                      style: const TextStyle(
+                        fontWeight: FontWeight.bold,
+                        color: Colors.blue,
+                      ),
+                    ),
+                  ],
                 ),
-                trailing: Text(
-                  '${data.duration?.inMilliseconds} ms',
+                trailing: Column(
+                  crossAxisAlignment: CrossAxisAlignment.end,
+                  mainAxisAlignment: MainAxisAlignment.center,
+                  children: [
+                    Text(
+                      '${data.duration?.inMilliseconds}ms',
+                    ),
+                    Row(
+                      mainAxisSize: MainAxisSize.min,
+                      children: [
+                        Text(
+                          data.statusCode.toString(),
+                          style: TextStyle(
+                            color: success ? Colors.green : Colors.red,
+                            fontWeight: FontWeight.bold,
+                            fontSize: 16,
+                          ),
+                        ),
+                      ],
+                    ),
+                  ],
                 ),
                 title: Text(
-                  data.url?.host ?? '<null>',
+                  '${data.url?.host}',
                   maxLines: 1,
                   overflow: TextOverflow.ellipsis,
                 ),
-                subtitle: Column(
-                  crossAxisAlignment: CrossAxisAlignment.start,
-                  children: [Text(data.url?.path ?? '<null>')],
-                ),
+                subtitle: Text(data.url?.path ?? noValue),
               );
             },
           );
@@ -91,33 +136,6 @@ class PlainRow extends StatelessWidget {
             ),
           )
       ],
-    );
-  }
-}
-
-class _LeadingStatusIncdicatorBlock extends StatelessWidget {
-  const _LeadingStatusIncdicatorBlock({
-    this.success = false,
-    this.method = '',
-  });
-
-  final bool success;
-  final String method;
-
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      padding: const EdgeInsets.all(6),
-      width: 60,
-      decoration: BoxDecoration(
-        color: success ? Colors.green : Colors.red,
-        borderRadius: BorderRadius.circular(3),
-      ),
-      child: Text(
-        method,
-        style: const TextStyle(color: Colors.white, fontSize: 10),
-        textAlign: TextAlign.center,
-      ),
     );
   }
 }